+++ src/dialogs/filedialog.h
@@ -1,0 +1,23 @@
+#ifndef FILEDIALOG_H
+#define FILEDIALOG_H
+
+
+#include <QtWidgets/QFileDialog>
+
+class FileDialog : public QFileDialog
+{
+Q_OBJECT
+
+public:
+    explicit FileDialog(QString name = "");
+    QString selectedFile();
+
+private:
+    QString _settingsKey;
+    QString _generalSettingsKey;
+
+protected slots:
+    void storeDirectory(QString path);
+};
+
+#endif // FILEDIALOG_H